Aggiunto un UncaughtExceptionHandler al client così da poter chiudere la connessione in caso di eccezioni non gestite, riposizionati gli observer del carrello e aggiornato il codice dei parsing all'interno di CartFragment per getire meglio eccezioni legate a problemi di parsing. Passaggio a BigDecimal.

+++ Client/app/src/main/java/com/example/cocktailapp/Bevanda.java
@@ -1,14 +1,15 @@
 package com.example.cocktailapp;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 public class Bevanda {
     private String nome;
-    private float prezzo;
+    private BigDecimal prezzo;
     private List<String> ingredienti;
     private int quantita;
 
-    public Bevanda(String nome, float prezzo, List<String> ingredienti, int quantita) {
+    public Bevanda(String nome, BigDecimal prezzo, List<String> ingredienti, int quantita) {
         this.nome = nome;
         this.prezzo = prezzo;
         this.ingredienti = ingredienti;
@@ -26,11 +27,11 @@ public class Bevanda {
         this.nome = nome;
     }
 
-    public float getPrezzo() {
+    public BigDecimal getPrezzo() {
         return prezzo;
     }
 
-    public void setPrezzo(float prezzo) {
+    public void setPrezzo(BigDecimal prezzo) {
         this.prezzo = prezzo;
     }
